fix bug where simultaneous dynamodb and kinesis event source listeners would clobber each other. fixed bug where only one listener could be active per stream. minor test reorganization.
