+++ tests/integration/test_dynamodb.py
@@ -1081,7 +1081,7 @@ class TestDynamoDB:
         events = retry(
             check_expected_lambda_log_events_length,
             retries=10,
-            sleep=3,
+            sleep=1,
             function_name=function_name,
             expected_length=1,
             regex_filter=r"Records",
